Move cluster span annotation into `package vtadminproto` to avoid circular import, trace vtsql
